@@ -22,64 +22,93 @@
   </modules>
 
   <properties>
-    <java.version>1.8</java.version>
+    <java.version>17</java.version>
     <maven.compiler.source>${java.version}</maven.compiler.source>
     <maven.compiler.target>${java.version}</maven.compiler.target>
     <maven.compiler.encoding>UTF-8</maven.compiler.encoding>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
+    <!---->
+    <!--继承父依赖不可修改版本号-->
     <!--SpringBoot-->
     <!--org.springframework.boot:spring-boot-starter-->
     <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter -->
-    <springboot.version>2.7.17</springboot.version>
-    <!--Maven编译插件-->
-    <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-compiler-plugin -->
-    <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
+    <springboot.version>3.3.1</springboot.version>
+
+    <!--SpringBoot Web-->
+    <!--org.springframework.boot:spring-boot-starter-web-->
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-web -->
+
+    <!--SpringBoot WebSocket-->
+    <!--org.springframework.boot:spring-boot-starter-websocket-->
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-web -->
+
+    <!--自动装配-->
+    <!--org.springframework.boot:spring-boot-autoconfigure-processor-->
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-autoconfigure-processor -->
+
+    <!--配置-->
+    <!--org.springframework.boot:spring-boot-configuration-processor-->
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-configuration-processor -->
+
+
+    <!---->
+    <!--继承父依赖可修改版本号-->
     <!--日志-->
     <!-- https://mvnrepository.com/artifact/ch.qos.logback/logback-classic -->
-    <logback.version>1.2.12</logback.version>
+    <logback.version>1.5.6</logback.version>
+
+    <!--Maven编译插件-->
+    <!-- https://mvnrepository.com/artifact/org.apache.maven.plugins/maven-compiler-plugin -->
+    <maven-compiler-plugin.version>3.13.0</maven-compiler-plugin.version>
+
+    <!--MongoDB-->
+    <!--org.springframework.boot:spring-boot-starter-data-mongodb-->
+    <!-- https://mvnrepository.com/artifact/org.springframework.boot/spring-boot-starter-data-mongodb -->
+    <!--<mongodb.version></mongodb.version>-->
+
+
+    <!---->
+    <!--父依赖不含-->
     <!--FastJson-->
     <!--com.alibaba.fastjson2:fastjson2-->
     <!-- https://mvnrepository.com/artifact/com.alibaba.fastjson2/fastjson2 -->
-    <fastjson.version>2.0.44</fastjson.version>
+    <fastjson.version>2.0.52</fastjson.version>
+
     <!--MQTT-->
     <!--org.eclipse.paho:org.eclipse.paho.client.mqttv3-->
     <!-- https://mvnrepository.com/artifact/org.eclipse.paho/org.eclipse.paho.client.mqttv3 -->
     <mqtt.version>1.2.5</mqtt.version>
+
     <!--加密算法-->
     <!--org.bouncycastle:bcpkix-jdk18on-->
     <!-- https://mvnrepository.com/artifact/org.bouncycastle/bcpkix-jdk18on -->
-    <bcpkix.version>1.77</bcpkix.version>
+    <bcpkix.version>1.78.1</bcpkix.version>
 
     <!--WebSocket SpringBoot模板-->
     <!--cn.404z:websocket-spring-boot-temp-->
     <websocket-spring-boot-temp.version>1.0.0</websocket-spring-boot-temp.version>
+
     <!--MongoDB SpringBoot启动器-->
     <!--cn.404z:mongo-spring-boot-starter-->
     <mongo-spring-boot-starter.version>1.0.0</mongo-spring-boot-starter.version>
+
     <!--MQTT SpringBoot启动器-->
     <!--cn.404z:mqtt-spring-boot-starter-->
     <mqtt-spring-boot-starter.version>1.0.0</mqtt-spring-boot-starter.version>
+
     <!--工具-->
     <!--cn.404z:util-->
     <util.version>1.0.0</util.version>
+
+    <!--SpringBoot工具-->
+    <!--cn.404z:util-spring-boot-->
+    <util-spring-boot.version>1.0.0</util-spring-boot.version>
   </properties>
 
   <dependencyManagement>
     <dependencies>
-      <!--日志-->
-      <dependency>
-        <groupId>ch.qos.logback</groupId>
-        <artifactId>logback-classic</artifactId>
-        <version>${logback.version}</version>
-      </dependency>
-      <!--FastJson-->
-      <dependency>
-        <groupId>com.alibaba.fastjson2</groupId>
-        <artifactId>fastjson2</artifactId>
-        <version>${fastjson.version}</version>
-      </dependency>
       <!--SpringBoot-->
       <dependency>
         <groupId>org.springframework.boot</groupId>
